update comments/move grid_parameters to grid
