+++ include/exadg/poisson/user_interface/application_base.h
@@ -37,7 +37,7 @@
 #include <exadg/poisson/user_interface/field_functions.h>
 #include <exadg/poisson/user_interface/parameters.h>
 #include <exadg/utilities/exceptions.h>
-#include <exadg/utilities/grid_parameters.h>
+#include <exadg/grid/grid_parameters.h>
 #include <exadg/utilities/output_parameters.h>
 #include <exadg/utilities/resolution_parameters.h>
